% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTo.R
\name{agghoo_run}
\alias{agghoo_run}
\title{agghoo_run}
\usage{
agghoo_run(dataTrain, dataTest, targetTrain, targetTest, floss, verbose, ...)
}
\description{
Run and eval the agghoo procedure.
Parameters are rather explicit except "floss", which corresponds to the
"final" loss function, applied to compute the error on testing dataset.
}
