% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_AgghooCV.R
\name{AgghooCV}
\alias{AgghooCV}
\title{R6 class with agghoo functions fit() and predict().}
\description{
Class encapsulating the methods to run to obtain the best predictor
from the list of models (see 'Model' class).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AgghooCV$new()}}
\item \href{#method-fit}{\code{AgghooCV$fit()}}
\item \href{#method-predict}{\code{AgghooCV$predict()}}
\item \href{#method-clone}{\code{AgghooCV$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new AgghooCV object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgghooCV$new(data, target, task, gmodel, quality = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Matrix or data.frame}

\item{\code{target}}{Vector of targets (generally numeric or factor)}

\item{\code{task}}{"regression" or "classification"}

\item{\code{gmodel}}{Generic model returning a predictive function}

\item{\code{quality}}{Function assessing the quality of a prediction;
quality(y1, y2) --> real number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit an agghoo model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgghooCV$fit(
  CV = list(type = "MC", V = 10, test_size = 0.2, shuffle = TRUE),
  mode = "agghoo"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CV}}{List describing cross-validation to run. Slots:
- type: 'vfold' or 'MC' for Monte-Carlo (default: MC)
- V: number of runs (default: 10)
- test_size: percentage of data in the test dataset, for MC
  (irrelevant for V-fold). Default: 0.2.
- shuffle: wether or not to shuffle data before V-fold.
  Irrelevant for Monte-Carlo; default: TRUE}

\item{\code{mode}}{"agghoo" or "standard" (for usual cross-validation)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict an agghoo model (after calling fit())
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgghooCV$predict(X, weight = "uniform")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Matrix or data.frame to predict}

\item{\code{weight}}{"uniform" (default) or "quality" to weight votes or
average models performances (TODO: bad idea?!)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgghooCV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
