% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTo.R
\name{compareRange}
\alias{compareRange}
\title{compareRange}
\usage{
compareRange(
  data,
  target,
  method_s,
  N = 100,
  nc = NA,
  floss = NULL,
  V_range = c(10, 15, 20),
  ...
)
}
\arguments{
\item{data}{Data matrix or data.frame}

\item{target}{Target vector (generally)}

\item{method_s}{Either a single function, or a list
(examples: agghoo_run, standardCV_run)}

\item{N}{Number of calls to method(s)}

\item{nc}{Number of cores. Set to parallel::detectCores() if undefined.
Set it to any value <=1 to say "no parallelism".}

\item{floss}{Loss function to compute the error on testing dataset.}

\item{V_range}{Values of V to be tested.}

\item{...}{arguments passed to method_s function(s)}
}
\description{
Run compareMulti on several values of the parameter V.
}
