% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTo.R
\name{compareTo}
\alias{compareTo}
\title{compareTo}
\usage{
compareTo(
  data,
  target,
  method_s,
  rseed = -1,
  floss = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{Data matrix or data.frame}

\item{target}{Target vector (generally)}

\item{method_s}{Either a single function, or a list
(examples: agghoo_run, standardCV_run)}

\item{rseed}{Seed of the random generator (-1 means "random seed")}

\item{floss}{Loss function to compute the error on testing dataset.}

\item{verbose}{TRUE to request methods to be verbose.}

\item{...}{arguments passed to method_s function(s)}
}
\description{
Compare a list of learning methods (or run only one), on data/target.
}
