% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agghoo.R
\name{compareToStandard}
\alias{compareToStandard}
\title{compareToStandard}
\usage{
compareToStandard(df, t_idx, task = NULL, rseed = -1)
}
\description{
Temporary function to compare agghoo to CV
(TODO: extended, in another file, more tests - when faster code).
}
