% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTo.R
\name{CVvoting_core}
\alias{CVvoting_core}
\title{CVvoting_core}
\usage{
CVvoting_core(data, target, task, gmodel, params, loss, CV)
}
\description{
"voting" cross-validation method, added here as an example.
Parameters are described in ?agghoo and ?AgghooCV
}
