% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTo.R
\name{CVvoting_run}
\alias{CVvoting_run}
\title{CVvoting_run}
\usage{
CVvoting_run(dataTrain, dataTest, targetTrain, targetTest, floss, verbose, ...)
}
\description{
Run and eval the voting cross-validation procedure.
Parameters are rather explicit except "floss", which corresponds to the
"final" loss function, applied to compute the error on testing dataset.
}
