% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTo.R
\name{standardCV_core}
\alias{standardCV_core}
\title{standardCV_core}
\usage{
standardCV_core(data, target, task, gmodel, params, loss, CV)
}
\description{
Cross-validation method, added here as an example.
Parameters are described in ?agghoo and ?AgghooCV
}
