% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTo.R
\name{standardCV_run}
\alias{standardCV_run}
\title{standardCV_run}
\usage{
standardCV_run(
  dataTrain,
  dataTest,
  targetTrain,
  targetTest,
  floss,
  verbose,
  ...
)
}
\description{
Run and eval the standard cross-validation procedure.
Parameters are rather explicit except "floss", which corresponds to the
"final" loss function, applied to compute the error on testing dataset.
}
