% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agghoo.R
\name{agghoo}
\alias{agghoo}
\title{agghoo}
\usage{
agghoo(data, target, task = NULL, gmodel = NULL, params = NULL, quality = NULL)
}
\arguments{
\item{data}{Data frame or matrix containing the data in lines.}

\item{target}{The target values to predict. Generally a vector.}

\item{task}{"classification" or "regression". Default:
regression if target is numerical, classification otherwise.}

\item{gmodel}{A "generic model", which is a function returning a predict
function (taking X as only argument) from the tuple
(dataHO, targetHO, param), where 'HO' stands for 'Hold-Out',
referring to cross-validation. Cross-validation is run on an array
of 'param's. See params argument. Default: see R6::Model.}

\item{params}{A list of parameters. Often, one list cell is just a
numerical value, but in general it could be of any type.
Default: see R6::Model.}

\item{quality}{A function assessing the quality of a prediction.
Arguments are y1 and y2 (comparing a prediction to known values).
Default: see R6::AgghooCV.}
}
\value{
An R6::AgghooCV object.
}
\description{
Run the agghoo procedure. (...)
}
\examples{
# Regression:
a_reg <- agghoo(iris[,-c(2,5)], iris[,2])
a_reg$fit()
pr <- a_reg$predict(iris[,-c(2,5)] + rnorm(450, sd=0.1))
# Classification
a_cla <- agghoo(iris[,-5], iris[,5])
a_cla$fit(mode="standard")
pc <- a_cla$predict(iris[,-5] + rnorm(600, sd=0.1))

}
