% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTo.R
\name{standardCV_run}
\alias{standardCV_run}
\title{standardCV_run}
\usage{
standardCV_run(
  dataTrain,
  dataTest,
  targetTrain,
  targetTest,
  floss,
  verbose,
  ...
)
}
\arguments{
\item{dataTrain}{Train dataset}

\item{dataTest}{Test dataset}

\item{targetTrain}{Train targets}

\item{targetTest}{Test targets}

\item{floss}{Loss function to compute error on test dataset}

\item{verbose}{Show some execution trace}

\item{...}{List defining the model (gmodel) and its parameters (params)}
}
\description{
Run and eval the standard cross-validation procedure.
}
